Replace SAM4L queue-based interrupts with NVIC regs
Instead of using a global queue to notify the event loop about pending
interrups, this commit uses the built in NVIC pending bits themselves.
This has a few benefits. First, it removes controller specific ISRs that
do nothing but enqueue an interrupt event. Second, it makes this code
generic for all Cortex-M4 chips (and should be ported to Cortex-M0 et
al). As a result of both of these, it removes about 10kB of code on hail
and imix. Third, it removes the need for controller drivers to handle
enabling/disabling nvic lines themselves. Fourth, it's much faster since
enqueing and dequeuing in the ring buffer is actually pretty expensive.