buttons: document and add active low/high
This commit adds the doc/syscalls/ documentation for the buttons driver.

It also adds an new enum `GpioMode` that allows boards to tell the
button driver if their button gpios will be low or high when the button
is pressed. This allows the driver to provide a consistent interface to
userland.

I decided that a button being pressed is a "1" and a button not pressed
is a "0" (as that seems intuitive). However, that is not how we have the
buttons wired on our platforms, and so the applications that were
relying on this undefined behavior were updated.